[Coroutines] Move Shape to its own header
* Plugin libraries to create custom ABIs will need access to CoroShape.
* As a step in enabling plugin libraries, move Shape into its own header
  that will eventually be moved into include/llvm/Transforms/Coroutines
